rename upgrade version
