+++ app/upgrades/v3_rc3/constants.go
@@ -1,7 +1,6 @@
-package v3_mainnet
+package v3_rc3
 
 import (
-	ibcfeetypes "github.com/cosmos/ibc-go/v8/modules/apps/29-fee/types"
 	ibctransfertypes "github.com/cosmos/ibc-go/v8/modules/apps/transfer/types"
 
 	storetypes "cosmossdk.io/store/types"
@@ -11,26 +10,17 @@ import (
 	vestingtypes "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 	"github.com/cosmos/cosmos-sdk/x/authz"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	consensusparamtypes "github.com/cosmos/cosmos-sdk/x/consensus/types"
-	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	govv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	"github.com/bandprotocol/chain/v3/app/upgrades"
-	bandtsstypes "github.com/bandprotocol/chain/v3/x/bandtss/types"
-	feedstypes "github.com/bandprotocol/chain/v3/x/feeds/types"
-	globalfeetypes "github.com/bandprotocol/chain/v3/x/globalfee/types"
 	oracletypes "github.com/bandprotocol/chain/v3/x/oracle/types"
-	restaketypes "github.com/bandprotocol/chain/v3/x/restake/types"
-	rollingseedtypes "github.com/bandprotocol/chain/v3/x/rollingseed/types"
-	tsstypes "github.com/bandprotocol/chain/v3/x/tss/types"
-	tunneltypes "github.com/bandprotocol/chain/v3/x/tunnel/types"
 )
 
 // UpgradeName defines the on-chain upgrade name.
 const (
-	UpgradeName = "v3_mainnet"
+	UpgradeName = "v3.0.0-rc3"
 
 	// BlockMaxBytes is the max bytes for a block, 3mb
 	BlockMaxBytes = int64(3000000)
@@ -42,20 +32,7 @@ const (
 var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,
-	StoreUpgrades: storetypes.StoreUpgrades{
-		Added: []string{
-			globalfeetypes.StoreKey,
-			consensusparamtypes.StoreKey,
-			crisistypes.StoreKey,
-			ibcfeetypes.StoreKey,
-			restaketypes.StoreKey,
-			feedstypes.StoreKey,
-			rollingseedtypes.StoreKey,
-			bandtsstypes.StoreKey,
-			tsstypes.StoreKey,
-			tunneltypes.StoreKey,
-		},
-	},
+	StoreUpgrades:        storetypes.StoreUpgrades{},
 }
 
 // TODO: Update ICA Allow messages
