+++ app/upgrades/v3_rc3/upgrades.go
@@ -1,4 +1,4 @@
-package v3
+package v3_rc3
 
 import (
 	"context"
